passes: Centralize cse
