+++ devito/passes/clusters/__init__.py
@@ -1,4 +1,5 @@
 from .manipulation import *  # noqa
+from .cse import *  # noqa
 from .rewriters import *  # noqa
 from .aliases import *  # noqa
 from .transformer import *  # noqa
